Move fetching utilities to entrypoint files.
